Created some components for using Imgui

+++ Common/Math/Vec2.cpp
@@ -2,7 +2,7 @@
 
 namespace Common
 {
-    Vec2::Vec2(int aX, int aY): x(aX), y(aY)
+    Vec2::Vec2(float aX, float aY): x(aX), y(aY)
     {
     }
 } // Common
